BackEnd/requirements.txt
fastapi [Standard]
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
pydantic-settings
